+++ test/functional-indexes.js
@@ -22,6 +22,7 @@ describe('indexes-tests', function () {
     services:{
       data:{
         config:{
+          autoUpdateDBVersion: true,
           datastores:[
             {
               name:'mongo',
@@ -42,6 +43,7 @@ describe('indexes-tests', function () {
     services:{
       data:{
         config:{
+          autoUpdateDBVersion: true,
           datastores:[
             {
               name:'mongo',
@@ -196,14 +198,25 @@ describe('indexes-tests', function () {
     var  mongodb = require('mongodb')
       ,  mongoclient = mongodb.MongoClient;
 
-    mongoclient.connect ("mongodb://127.0.0.1:27017", { database: "indexes_configured_test_db" }, function (err, database) {
+    mongoclient.connect ("mongodb://127.0.0.1:27017", function (err, client) {
 
       if (err) return callback(err);
 
-      var collection = database.collection("happn");
+      var database = client.db("indexes_configured_test_db");
+
+      var collection = database.collection("indexes_configured_test_db_coll");
 
       collection.listIndexes().toArray(function(e, indexes){
-        expect(indexes).to.eql([{"v":1,"key":{"_id":1},"ns":"admin.happn","name":"_id_"},{"v":1,"key":{"happn_path_index":1},"ns":"admin.happn","name":"happn_path_index_1","path":1},{"v":1,"key":{"path":1},"unique":true,"ns":"admin.happn","name":"path_1"},{"v":1,"key":{"test":1},"ns":"admin.happn","name":"test_1"}]);
+
+        var uniqueFound = false;
+
+        for (var i = 0; i < indexes.length; i++){
+
+          var index = indexes[i];
+          if (index["unique"] && index["key"] && index["key"].path == 1) uniqueFound = true;
+        }
+
+        expect(uniqueFound).to.be(true);
         done();
       });
     });