+++ test/functional_happn.js
@@ -18,6 +18,7 @@ describe('happn-tests', function () {
     services:{
       data:{
         config:{
+          autoUpdateDBVersion: true,
           datastores:[
             {
               name:'mongo',
@@ -1158,24 +1159,5 @@ describe('happn-tests', function () {
         expect(e).to.not.be.ok();
       });
     });
-
   });
-
-  it('will do events in the order they are passed', function (done) {
-    publisherclient.set('/test_event_order', {property1: 'property1Value'}, {}, function () {
-      publisherclient.log.info('Done setting');
-    });
-    publisherclient.remove('/test_event_order', function (err) {
-      publisherclient.log.info('Done removing');
-      setTimeout(function () {
-        publisherclient.get('/test_event_order', null, function (e, result) {
-          expect(result).to.be(null);
-          done();
-        });
-      }, 1000);
-    });
-  });
-
-  //require('benchmarket').stop();
-
 });